custom: add workaround for vxlan restart bug
When the vx0 interface is a child of a DSA bridge or a wireless bridge and the parent device restarts or halts, vx0 does not restore to proper connectivity.
This means a user has to manual restart vx0 for proper connectivity, in rare cases both ends go down.

as temporary fix to what likely is caused by some type of race condition in netifd or a unawareness,
lets just restart vx0 with a hotplug event every 5 seconds it reports a ifdown event.