hostapd: add SAE support for wifi-station
Important Note: unlike PSK wifi-stations, both mac & key are required
to make it work. With PSK, hostapd used to do a bruteforce match to
find which PSK entry to use but with SAE this is infeasible due to
SAE's design.

When mac is omitted, it will allow any mac address to use the SAE
password if it didn't have a mac address assigned to it but this
could only be done ONCE. The last wildcard entry would be used.

Also unlike "hostapd: add support for SAE in PPSK option" (913368a),
it is not required to set sae_pwe to 0. This gives it a slight
advantage over using PPSK that goes beyond not needing RADIUS.

Example Configuration:

```
config wifi-vlan
        option iface default_radio0
        option name 999
        option vid 999
        option network management

config wifi-station
	# Allow user with mac address 00:11:22:33:44:55 and matching
	# key "secretadminpass" to access the management network.
        option iface default_radio0
	option vid 999
	option mac '00:11:22:33:44:55'
        option key secretadminpass

config wifi-vlan
        option iface default_radio0
        option name 100
        option vid 100
        option network guest

config wifi-station
	# With SAE, when mac is omitted it will be the fallback in case no
	# other mac address matches. It won't be possible for a user that
	# has a matching mac to use this network (i.e., 00:11:22:33:44:55
	# in this example).
        option iface default_radio0
	option vid 100
        option key guestpass
```

Signed-off-by: Rany Hany <rany_hany@riseup.net>